feat: change types and add other validations on strikers

+++ src/controllers/StrikerController.ts
@@ -8,7 +8,7 @@ import generateArray from '../utils/generateArray';
 const { Striker } = db;
 
 export default class StrikerController {
-  static async insert(req: Request, res: Response): Promise<Response | void> {
+  static async insert(req: Request, res: Response): Promise<void> {
     try {
       const topScorer: topScorerInterface[] = req.body.topScorer;
 
@@ -23,32 +23,42 @@ export default class StrikerController {
     }
   }
 
-  static async show(req: Request, res: Response): Promise<Response> {
+  static async show(req: Request, res: Response): Promise<void> {
     try {
-      const strikers = await Striker.findAll();
+      const strikers: StrikerAttributes = await Striker.findAll();
 
-      return res.send(strikers);
+      res.send(strikers);
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 
-  static async index(req: Request, res: Response): Promise<Response> {
+  static async index(req: Request, res: Response): Promise<Response | void> {
     const { id } = req.params;
 
     try {
-      const striker = await Striker.findByPk(id);
+      const striker: StrikerAttributes = await Striker.findByPk(id);
 
-      return res.send(striker);
+      if (!striker) {
+        return res.status(204).send();
+      }
+
+      res.send(striker);
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 
   static async update(req: Request, res: Response): Promise<Response | void> {
-    const { id } = req.params;
     try {
+      const { id } = req.params;
       const topScorer: topScorerInterface[] = req.body.topScorer;
+
+      const striker: StrikerAttributes = await Striker.findByPk(id);
+
+      if (!striker) {
+        return res.status(400).send({ message: 'striker not found' });
+      }
 
       const valid = customValidations(topScorer, res);
 
@@ -56,36 +66,41 @@ export default class StrikerController {
 
       await Striker.update(req.body, { where: { id } });
 
-      const striker = await Striker.findByPk(id);
+      const newStriker = await Striker.findByPk(id);
 
-      return res.send(striker);
+      res.send(newStriker);
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 
-  static async delete(req: Request, res: Response): Promise<Response> {
-    const { id } = req.params;
-
+  static async delete(req: Request, res: Response): Promise<Response | void> {
     try {
+      const { id } = req.params;
+      const striker: StrikerAttributes = await Striker.findByPk(id);
+
+      if (!striker) {
+        return res.status(400).send({ message: 'striker not found' });
+      }
+
       await Striker.destroy({ where: { id } });
 
-      return res.send({ message: 'Successfully delete striker' });
+      res.send({ message: 'Successfully delete striker' });
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 
   static async teamWithMoreStrikers(
     req: Request,
     res: Response
-  ): Promise<Response> {
+  ): Promise<void> {
     try {
-      const findStrikers = await Striker.findAll();
+      const findStrikers: StrikerAttributes[] = await Striker.findAll();
 
-      const strikers: Object[] = [];
+      const strikers: Array<StrikerAttributes> = [];
 
-      findStrikers.map((element: StrikerAttributes) => {
+      findStrikers.map((element: any) => {
         strikers.push(...element.topScorer);
       });
 
@@ -93,13 +108,13 @@ export default class StrikerController {
 
       const team = orderStrikers.slice(0, 1);
 
-      return res.send(...team);
+      res.send(...team);
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 
-  static async topStrikers(req: Request, res: Response): Promise<Response> {
+  static async topStrikers(req: Request, res: Response): Promise<void> {
     try {
       const findStrikers: Object[] = await Striker.findAll({
         order: [['goals', 'DESC']],
@@ -108,15 +123,18 @@ export default class StrikerController {
 
       const topStrikers = findStrikers.slice(0, 5);
 
-      return res.send(topStrikers);
+      res.send(topStrikers);
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 
-  static async listByGoals(req: Request, res: Response): Promise<Response> {
+  static async listByGoals(
+    req: Request,
+    res: Response
+  ): Promise<Response | void> {
     try {
-      const { goals } = req.body;
+      const goals: number = req.body.goals;
 
       if (goals < 1)
         return res.status(400).send({ error: 'invalid number of goals' });
@@ -128,9 +146,9 @@ export default class StrikerController {
 
       if (strikers.length === 0) return res.status(204).send();
 
-      return res.send(strikers);
+      res.send(strikers);
     } catch (err) {
-      return res.status(400).send({ err });
+      res.status(400).send({ err });
     }
   }
 }
